+++ src/function/preset.ts
@@ -6,7 +6,7 @@ import { oai_settings, promptManager } from '@sillytavern/scripts/openai';
 import { uuidv4 } from '@sillytavern/scripts/utils';
 import { LiteralUnion, PartialDeep, SetRequired } from 'type-fest';
 
-type Preset = {
+export type Preset = {
   settings: {
     max_context: number;
     max_completion_tokens: number;
@@ -53,7 +53,7 @@ type Preset = {
   };
 };
 
-type PresetPrompt = {
+export type PresetPrompt = {
   id: LiteralUnion<
     | 'main'
     | 'nsfw'
@@ -85,12 +85,12 @@ type PresetPrompt = {
 
   extra?: Record<string, any>;
 };
-type PresetNormalPrompt = SetRequired<{ id: string } & Omit<PresetPrompt, 'id'>, 'position' | 'content'>;
-type PresetSystemPrompt = SetRequired<
+export type PresetNormalPrompt = SetRequired<{ id: string } & Omit<PresetPrompt, 'id'>, 'position' | 'content'>;
+export type PresetSystemPrompt = SetRequired<
   { id: 'main' | 'nsfw' | 'jailbreak' | 'enhanceDefinitions' } & Omit<PresetPrompt, 'id'>,
   'content'
 >;
-type PresetPlaceholderPrompt = SetRequired<
+export type PresetPlaceholderPrompt = SetRequired<
   {
     id:
       | 'worldInfoBefore'