+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 <!-- markdownlint-disable MD041 MD036 -->
+## 4.8.7
+
+### 📦函数
+
+- `generate` 现在支持选择其他预设了, 如 `generate({ preset_name: '其他预设名称' })`; 但只会使用所选预设的提示词及参数, 不会使用所选预设的正则、酒馆助手脚本
+
 ## 4.8.6
 
 ### 🗣提示词查看器
@@ -30,7 +36,7 @@
 ### 📦函数
 
 - 新增 `getProxyPresetNames` 函数, 用于获取代理预设名称列表
-- `generate`和`generateRaw`函数现在支持工具调用, 支持的源详见[文档](https://n0vi028.github.io/JS-Slash-Runner-Doc/guide/功能详情/请求生成.html#custom-api)支持的 API 源部分 (由GitHub用户 @ultranationalism 贡献)
+- `generate` 和 `generateRaw` 函数现在支持工具调用, 支持的源详见[文档](https://n0vi028.github.io/JS-Slash-Runner-Doc/guide/功能详情/请求生成.html#custom-api)支持的 API 源部分 (由GitHub用户 @ultranationalism 贡献)
 
 ## 4.8.2